refactor: huge refactor, moving validations to use cases

+++ src/interface-adapters/controllers/user/delete-user-by-id-controller.ts
@@ -1,21 +1,23 @@
+import { RequestValidationError } from '~/application/errors/request-validation-error';
 import { Controller } from '~/application/ports/controllers/controller';
 import { Presenter } from '~/application/ports/presenters/presenter';
 import { RequestModel } from '~/application/ports/requests/request-model';
 import { ResponseModel } from '~/application/ports/responses/response-model';
-import { DeleteUserByIdUseCase } from '~/domain/user/use-cases/delete-user-by-id-use-case';
-import { ValidationComposite } from '~/application/ports/validation/validation-composite';
+import { DeleteUserByIdUseCase } from '~/application/ports/use-cases/user/delete-user-by-id-use-case';
 
-export class DeleteUserByIdController implements Controller<void> {
+export class DeleteUserByIdController implements Controller<void | never> {
   constructor(
     private readonly deleteUserByIdUseCase: DeleteUserByIdUseCase,
-    private readonly validation: ValidationComposite,
     private readonly presenter: Presenter<void>,
   ) {}
 
   async handleRequest(
     requestModel: RequestModel,
-  ): Promise<ResponseModel<void>> {
-    await this.validation.validate(requestModel);
+  ): Promise<ResponseModel<void>> | never {
+    if (!requestModel || !requestModel.params || !requestModel.params.id) {
+      throw new RequestValidationError('Missing params');
+    }
+
     await this.deleteUserByIdUseCase.deleteById(requestModel.params.id);
     return await this.presenter.response();
   }
